Moved marker components to common folder to being able to use it for scatter plots as well.
