+++ src/common/marker/VoronoiCircle.jsx
@@ -2,17 +2,16 @@
 
 var React = require('react');
 var d3 = require('d3');
-var shade = require('../utils').shade;
+var shade = require('../../utils/index').shade;
 var VoronoiArea = require('./VoronoiArea');
 
 module.exports = React.createClass({
 
-  displayName: 'VoronoiRect',
+  displayName: 'VoronoiCircle',
 
   getDefaultProps() {
     return { 
-      markerWidth: 6,
-      markerHeight: 6,
+      markerRadius: 3,
       markerFill: '#1f77b4',
       hoverAnimation: true
     };
@@ -20,8 +19,7 @@ module.exports = React.createClass({
 
   getInitialState() {
     return {
-      markerWidth: this.props.markerWidth,
-      markerHeight: this.props.markerHeight,
+      markerRadius: this.props.markerRadius,
       markerFill: this.props.markerFill
     }
   },
@@ -43,15 +41,14 @@ module.exports = React.createClass({
           handleMouseLeave={handleMouseLeave}
           voronoiPath={this.props.voronoiPath}
         />
-        <rect
+        <circle
           onMouseOver={handleMouseOver}
           onMouseLeave={handleMouseLeave}
-          x={this.props.cx - Math.round(this.props.markerWidth/2)}
-          y={this.props.cy - Math.round(this.props.markerHeight/2)}
-          width={this.state.markerWidth}
-          height={this.state.markerHeight}
+          cx={this.props.cx}
+          cy={this.props.cy}
+          r={this.state.markerRadius}
           fill={this.state.markerFill}
-          className="rd3-linechart-rect"
+          className="rd3-linechart-circle"
         />
       </g>
     );
@@ -59,15 +56,14 @@ module.exports = React.createClass({
 
   _animateMarker() {
     this.setState({
-      markerWidth: this.props.markerWidth * ( 5 / 4 ),
-      markerHeight: this.props.markerHeight * ( 5 / 4 ),
+      markerRadius: this.props.markerRadius * ( 5 / 4 ),
       markerFill: shade(this.props.markerFill, 0.2)
     });
   },
 
   _restoreMarker() {
     this.setState(
-        this.getInitialState()
+      this.getInitialState()
     );
   }
 });
