+++ src/common/marker/VoronoiRect.jsx
@@ -2,16 +2,17 @@
 
 var React = require('react');
 var d3 = require('d3');
-var shade = require('../utils').shade;
+var shade = require('../../utils/index').shade;
 var VoronoiArea = require('./VoronoiArea');
 
 module.exports = React.createClass({
 
-  displayName: 'VoronoiCircle',
+  displayName: 'VoronoiRect',
 
   getDefaultProps() {
     return { 
-      markerRadius: 3,
+      markerWidth: 6,
+      markerHeight: 6,
       markerFill: '#1f77b4',
       hoverAnimation: true
     };
@@ -19,7 +20,8 @@ module.exports = React.createClass({
 
   getInitialState() {
     return {
-      markerRadius: this.props.markerRadius,
+      markerWidth: this.props.markerWidth,
+      markerHeight: this.props.markerHeight,
       markerFill: this.props.markerFill
     }
   },
@@ -41,14 +43,15 @@ module.exports = React.createClass({
           handleMouseLeave={handleMouseLeave}
           voronoiPath={this.props.voronoiPath}
         />
-        <circle
+        <rect
           onMouseOver={handleMouseOver}
           onMouseLeave={handleMouseLeave}
-          cx={this.props.cx}
-          cy={this.props.cy}
-          r={this.state.markerRadius}
+          x={this.props.cx - Math.round(this.props.markerWidth/2)}
+          y={this.props.cy - Math.round(this.props.markerHeight/2)}
+          width={this.state.markerWidth}
+          height={this.state.markerHeight}
           fill={this.state.markerFill}
-          className="rd3-linechart-circle"
+          className="rd3-linechart-rect"
         />
       </g>
     );
@@ -56,14 +59,15 @@ module.exports = React.createClass({
 
   _animateMarker() {
     this.setState({
-      markerRadius: this.props.markerRadius * ( 5 / 4 ),
+      markerWidth: this.props.markerWidth * ( 5 / 4 ),
+      markerHeight: this.props.markerHeight * ( 5 / 4 ),
       markerFill: shade(this.props.markerFill, 0.2)
     });
   },
 
   _restoreMarker() {
     this.setState(
-      this.getInitialState()
+        this.getInitialState()
     );
   }
 });
