+++ src/common/marker/index.js
@@ -1,0 +1,11 @@
+/**
+ * Created by cfaes on 24.08.2015.
+ */
+// general
+exports.VoronoiArea = require('./VoronoiArea');
+exports.VoronoiContainer = require('./VoronoiContainer');
+
+// markers
+exports.VoronoiCircle = require('./VoronoiCircle');
+exports.VoronoiRect = require('./VoronoiRect');
+exports.VoronoiStar = require('./VoronoiStar');
